[SPARK-41399][CONNECT] Refactor column related tests to test_connect_column
### What changes were proposed in this pull request?

Given there is a dedicated `test_connect_column.py` now, we should move those column API tests to this place.

### Why are the changes needed?

Codebase refactoring

### Does this PR introduce _any_ user-facing change?

NO

### How was this patch tested?

Existing UT

Closes #38925 from amaliujia/move_all_column_tests_to_right_place.

Authored-by: Rui Wang <rui.wang@databricks.com>
